fix macos build error

+++ libs/containers/include/mrpt/containers/yaml.h
@@ -107,9 +107,9 @@ class yaml
 		~node_t() = default;
 
 		template <
-			typename T,	 //
+			typename T,  //
 			typename = std::enable_if_t<!std::is_constructible_v<
-				std::initializer_list<map_t::value_type>, T>>,	//
+				std::initializer_list<map_t::value_type>, T>>,  //
 			typename = std::enable_if_t<!std::is_constructible_v<
 				std::initializer_list<sequence_t::value_type>, T>>>
 		node_t(const T& scalar)
@@ -966,8 +966,8 @@ template <typename MATRIX>
 inline yaml yaml::FromMatrix(const MATRIX& m)
 {
 	yaml r = mrpt::containers::yaml::Map();
-	r["rows"] = m.rows();
-	r["cols"] = m.cols();
+	r["rows"] = static_cast<int64_t>(m.rows());
+	r["cols"] = static_cast<int64_t>(m.cols());
 	auto& data = r["data"] = mrpt::containers::yaml::Sequence();
 	data.node().printInShortFormat = true;
 	for (int iRow = 0; iRow < m.rows(); iRow++)
